requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==20.3.0
autopep8==1.6.0
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.3
click==8.1.3
colorama==0.4.4
cookiecutter==1.7.3
cryptography==36.0.2
Django==4.2.3
django-browser-reload==1.10.0
django-extensions==3.2.3
django-field-history==0.8.0
django-model-utils==4.3.1
django-stubs==4.2.1
django-stubs-ext==4.2.1
docopt==0.6.2
frozenlist==1.3.3
graphviz==0.20.1
idna==3.2
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
keyboard==0.13.5
lib50==3.0.4
libsass==0.21.0
markdown2==2.4.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
multidict==6.0.4
mypy==1.4.0
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.8
opencv-python==4.7.0.68
pexpect==4.8.0
Pillow==8.3.2
pipreqs==0.4.11
poyo==0.5.0
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
PyGetWindow==0.0.9
pygraphviz==1.11
PyMsgBox==1.0.9
pyparsing==3.1.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.2
python-slugify==6.1.2
PyTweening==1.0.3
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
submit50==3.1.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing_extensions==4.6.3
tzdata==2022.1
urllib3==1.26.6
yarg==0.1.9
yarl==1.9.2
